Improves progress reporting for public data transfer
Adds job status updates during the transfer of public
data to provide more granular feedback on the
progress of the operation. This includes updates
at various stages of the file processing and transfer,
allowing for better monitoring and understanding
of the task's advancement. Also corrects an error
where the gearman worker object was not correctly
passed to the transfer function.