Improves PublicData transfer reliability
Adds logging of excluded files during the PublicData transfer process,
which aids in debugging and identifying potential issues with file
transfers. The excluded files are now written to a dedicated log file.

Removes the `remove_source_files` argument from the
`transfer_publicdata_dir` function and stops attempting to clear out the
public data directory as part of the cruise finalization.