Adds directory exemption to lockdown function
Extends `lockdown_directory` to accept a list of directories
that should be excluded from permission changes. This is
necessary to allow the cruise directory to remain accessible
while still restricting access to the parent directory. Also
adds error handling to the lockdown function and returns error
messages to the calling functions.